Ayyanar commit on fixed after validation

+++ backend/database/migrations/2024_04_12_191646_create_user_work_history_table.php
@@ -22,7 +22,7 @@ class CreateUserWorkHistoryTable extends Migration
             $table->string('supervisor_name');
             $table->string('employer_email');
             $table->string('employer_fax');
-            $table->integer('employer_phone');
+            $table->bigInteger('employer_phone');
             $table->timestamps();            
         });
     }
