+++ backend/database/migrations/2024_04_12_191958_create_user_emergency_contacts_table.php
@@ -20,7 +20,7 @@ class CreateUserEmergencyContactsTable extends Migration
             $table->string('relationship');
             $table->string('relationship_name');
             $table->string('relationship_email');
-            $table->integer('relationship_phone');
+            $table->bigInteger('relationship_phone');
             $table->timestamps();
         });
     }
